refactor(tests): adopted unified plugin directory layout
