+++ packages/wp-plugin/ionos-essentials/ionos-essentials/inc/dashboard/tests/e2e/next-best-actions.spec.js
@@ -1,5 +1,5 @@
 import { test, expect } from '@wordpress/e2e-test-utils-playwright';
-import { execTestCLI } from '../../../../../../../playwright/wp-env.js';
+import { execTestCLI } from '../../../../../../../../playwright/wp-env.js';
 
 test.describe('essentials:dashboard next-best-actions block', () => {
   test.beforeAll(async () => {
