+++ packages/wp-plugin/ionos-essentials/ionos-essentials/inc/dashboard/tests/e2e/welcome.spec.js
@@ -1,12 +1,10 @@
 import { test, expect } from '@wordpress/e2e-test-utils-playwright';
-import { execTestCLI } from '../../../../../../../playwright/wp-env.js';
+import { execTestCLI } from '../../../../../../../../playwright/wp-env.js';
 
 test.describe('essentials:dashboard next-best-actions block', () => {
   test.beforeAll(async () => {
     try {
-      execTestCLI(
-        `wp user meta delete 1 ionos_essentials_welcome`
-      );
+      execTestCLI(`wp user meta delete 1 ionos_essentials_welcome`);
     } catch (error) {}
   });
