+++ packages/wp-plugin/ionos-essentials/ionos-essentials/inc/maintenance_mode/tests/e2e/maintenance.spec.js
@@ -1,20 +1,14 @@
 import { test, expect } from '@wordpress/e2e-test-utils-playwright';
-import { execTestCLI } from '../../../../../../../playwright/wp-env.js';
+import { execTestCLI } from '../../../../../../../../playwright/wp-env.js';
 
 test.describe('Maintenance', () => {
-
   test.beforeAll(async () => {
-      execTestCLI(
-        `wp option delete ionos_essentials_maintenance_mode`
-      );
+    execTestCLI(`wp option delete ionos_essentials_maintenance_mode`);
   });
 
   test.afterAll(async () => {
-      execTestCLI(
-        `wp option delete ionos_essentials_maintenance_mode`
-      );
+    execTestCLI(`wp option delete ionos_essentials_maintenance_mode`);
   });
-
 
   test('maintenance mode is enabled', async ({ admin, page }) => {
     await admin.visitAdminPage('/');
@@ -26,7 +20,6 @@ test.describe('Maintenance', () => {
     let body = await page.locator('body');
     await expect(body).toHaveText(/Blog/);
 
-
     await page.goto('/wp-login.php?action=logout');
     await page.click('text=log out');
 
@@ -35,6 +28,4 @@ test.describe('Maintenance', () => {
     await page.waitForTimeout(1000);
     await expect(page).toHaveTitle('Construction');
   });
-
-
 });
