+++ packages/wp-plugin/ionos-essentials/ionos-essentials/inc/security/tests/e2e/security.spec.js
@@ -1,6 +1,6 @@
 import { test, expect } from '@wordpress/e2e-test-utils-playwright';
 
-import { execTestCLI } from '../../../../../../../playwright/wp-env.js';
+import { execTestCLI } from '../../../../../../../../playwright/wp-env.js';
 
 test.beforeAll(async () => {
   try {
